UDP: Refactor some dissector code
Add UDP length as a generated field of UDP-Lite subtree.

Reviewed-on: https://code.wireshark.org/review/10775
Petri-Dish: Evan Huus <eapache@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
(cherry picked from commit 9fea515efcd025140d018df6d1cf4e3d9ee4f4ac)
Reviewed-on: https://code.wireshark.org/review/10853

+++ epan/dissectors/packet-udp.h
@@ -35,7 +35,7 @@ extern "C" {
 typedef struct _e_udphdr {
   guint16 uh_sport;
   guint16 uh_dport;
-  guint16 uh_ulen;
+  guint32 uh_ulen;
   guint32 uh_sum_cov;
   guint16 uh_sum;
   guint32 uh_stream; /* this stream index field is included to help differentiate when address/port pairs are reused */
